Float

+++ SU2_PY/SU2/run/amg.py
@@ -472,7 +472,7 @@ def amg ( config , kind='' ):
                 
                 # config_cfd.RESIDUAL_REDUCTION = float(adap_res[iSiz])
                 config_cfd.ITER               = int(adap_flow_iter[iSiz])
-                config_cfd.CFL_NUMBER         = int(adap_cfl[iSiz])
+                config_cfd.CFL_NUMBER         = float(adap_cfl[iSiz])
                 
                 config_cfd.WRT_BINARY_RESTART  = "NO"
                 config_cfd.READ_BINARY_RESTART = "NO"
